lowmemorykiller: maintain LMK rbtree with signal->adj_node
Currently, we maintain LMK rbtree with task->adj_node. However, when
handling oom_score_adj change case, we may del/add a non-leader task
to the RB tree, which is not as expected.

This patch we maintain the LMK rbtree with task->signal->adj_node.
Since signal_struct is shared between main task and threads, we can
avoid non-leader thread adding to tree.